+++ lib/features/aliases/presentation/pages/alias_page.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:practice_acount_manager/features/aliases/presentation/components/search_table_alias.dart';
 import 'package:practice_acount_manager/features/widgets/generals/button_aliase_navigation.dart';
-import 'package:practice_acount_manager/features/widgets/generals/button_user_navigation.dart';
 import 'package:practice_acount_manager/features/widgets/generals/drawer.dart';
 import 'package:practice_acount_manager/features/widgets/generals/footer.dart';
 import 'package:practice_acount_manager/l10n/app_localizations.dart';
@@ -48,21 +47,7 @@ class AliasPage extends StatelessWidget {
           ],
         ),
       ),
-      // body: SingleChildScrollView(
-      //   padding: const EdgeInsets.all(16.0),
-      //   child: Column(
-      //     crossAxisAlignment: CrossAxisAlignment.center,
-      //     children: [
-      //       Center(child: ButtonOptionsAliase()),
-      //       SizedBox(height: 20),
-
-      //       //SearchBarExample(),
-      //       SizedBox(child: SearchTableAliases()),
-      //     ],
-      //   ),
-      // ),
       bottomNavigationBar: const Footer(),
-      //floatingActionButton: const ButtonHome(),
       floatingActionButtonLocation: FloatingActionButtonLocation.endFloat,
     );
   }
