+++ lib/features/auth/presentation/service/auth_service.dart
@@ -2,14 +2,16 @@ import 'dart:convert';
 import 'dart:math';
 import 'package:crypto/crypto.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:http/http.dart' as http;
-import 'package:practice_acount_manager/features/auth/presentation/pages/select_login_page.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:app_links/app_links.dart';
 import 'package:flutter/material.dart';
-
 import 'dart:async';
+
+import 'package:practice_acount_manager/features/auth/presentation/pages/select_login_page.dart';
+import 'package:practice_acount_manager/riverpod/auth_provider.dart';
 
 class AuthService {
   final String zitadelUrl;
@@ -18,6 +20,7 @@ class AuthService {
 
   String? _codeVerifier;
   final AppLinks _appLinks = AppLinks();
+  final AuthServiceProvider = Provider<AuthService>((ref) => AuthService());
 
   StreamSubscription? _sub;
 
@@ -147,64 +150,21 @@ class AuthService {
       throw Exception('Error al obtener información del usuario');
     }
   }
+
+  @override
+  Widget build(BuildContext context) {
+    // TODO: implement build
+    throw UnimplementedError();
+  }
 }
 
-// Future<void> logout(BuildContext context, {bool isZitadel = false}) async {
-//   final prefs = await SharedPreferences.getInstance();
-
-//   if (isZitadel) {
-//     final idToken = prefs.getString('id_token') ?? '';
-//     const redirectUri = 'com.practiceacountmanager.app://callback';
-
-//     final logoutUrl = Uri.parse(
-//       'https://adminemail-prueba-ftulnf.us1.zitadel.cloud/oidc/v1/end_session'
-//       '?id_token_hint=$idToken'
-//       '&post_logout_redirect_uri=$redirectUri',
-//     );
-
-//     await prefs.clear();
-
-//     if (await canLaunchUrl(logoutUrl)) {
-//       await launchUrl(logoutUrl, mode: LaunchMode.externalApplication);
-//     } else {
-//       debugPrint('No se pudo abrir el navegador para cerrar sesión en Zitadel');
-//     }
-//   } else {
-//     // Logout local
-//     final token = prefs.getString('access_token') ?? '';
-//     final String apiService = dotenv.env['API_SERVICE'] ?? '';
-//     try {
-//       await http.post(
-//         Uri.parse('$apiService/user/logout'),
-//         headers: {
-//           'Accept': 'application/json',
-//           'Content-Type': 'application/json',
-//           'X-Client-Type': 'mobile',
-//           if (token.isNotEmpty) 'Authorization': 'Bearer $token',
-//         },
-//       );
-//     } catch (e) {
-//       debugPrint('Error logout local: $e');
-//     }
-
-//     await prefs.clear();
-//   }
-
-//   // Navegar a la pantalla de login eliminando historial
-//   if (context.mounted) {
-//     Navigator.of(context).pushAndRemoveUntil(
-//       MaterialPageRoute(builder: (_) => const SelectLoginPage()),
-//       (route) => false,
-//     );
-//   }
-// }
-
-Future<void> logout(BuildContext context) async {
+Future<void> logout(BuildContext context, WidgetRef ref) async {
   final prefs = await SharedPreferences.getInstance();
   final idToken = prefs.getString('id_token') ?? '';
   const redirectUri = 'com.practiceacountmanager.app://callback';
 
   await prefs.clear();
+  ref.read(authProvider.notifier).clearTokens();
 
   final logoutUrl = Uri.parse(
     'https://adminemail-prueba-ftulnf.us1.zitadel.cloud/oidc/v1/end_session'
