+++ lib/features/perfil/presentation/pages/informacion_perfil.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class InformacionPerfil extends StatefulWidget {
+  const InformacionPerfil({super.key});
+
+  @override
+  State<InformacionPerfil> createState() => _InformacionPerfilState();
+}
+
+class _InformacionPerfilState extends State<InformacionPerfil> {
+  @override
+  Widget build(BuildContext context) {
+    return Placeholder();
+  }
+}
