+++ lib/riverpod/auth_provider.dart
@@ -1,21 +1,29 @@
 import 'dart:async';
+import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:jwt_decode/jwt_decode.dart';
 import 'package:practice_acount_manager/features/auth/presentation/pages/select_login_page.dart';
 import 'package:practice_acount_manager/features/auth/presentation/service/auth_service_local.dart';
 import 'package:practice_acount_manager/features/core/navigation.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 class AuthState {
   final String? accessToken;
   final String? refreshToken;
+  final Map<String, dynamic>? userInfo;
 
-  AuthState({this.accessToken, this.refreshToken});
+  AuthState({this.accessToken, this.refreshToken, this.userInfo});
 
-  AuthState copyWith({String? accessToken, String? refreshToken}) {
+  AuthState copyWith({
+    String? accessToken,
+    String? refreshToken,
+    Map<String, dynamic>? userInfo,
+  }) {
     return AuthState(
       accessToken: accessToken ?? this.accessToken,
       refreshToken: refreshToken ?? this.refreshToken,
+      userInfo: userInfo ?? this.userInfo,
     );
   }
 }
@@ -25,6 +33,27 @@ class AuthNotifier extends StateNotifier<AuthState> {
   final LoginService _loginService = LoginService();
 
   AuthNotifier() : super(AuthState());
+
+  Future<void> loadTokens() async {
+    final prefs = await SharedPreferences.getInstance();
+
+    final access = prefs.getString('id_token') ?? '';
+    final refresh = prefs.getString('refresh_token') ?? '';
+
+    // Leer userInfo como string
+    final userJson = prefs.getString('user_info');
+    Map<String, dynamic>? userInfo;
+    if (userJson != null && userJson.isNotEmpty) {
+      userInfo = jsonDecode(userJson);
+    }
+
+    // Guardar en el state
+    state = AuthState(
+      accessToken: access,
+      refreshToken: refresh,
+      userInfo: userInfo,
+    );
+  }
 
   void setTokens({required String accessToken, required String refreshToken}) {
     state = state.copyWith(
